unoptimized support for Csc matrix, kaizhang/anndata-rs#5
