+++ anndata/tests/anndata_op.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code, unused)]
+
 mod common;
 use common::*;
 
@@ -5,6 +7,9 @@ use ndarray::Array2;
 use proptest::prelude::*;
 use anndata::*;
 use anndata_hdf5::H5;
+use std::path::Path;
+use nalgebra_sparse::{CscMatrix, CsrMatrix};
+
 
 fn test_speacial_cases<F, T>(adata_gen: F)
 where
@@ -116,4 +121,29 @@ fn test_iterator_h5() {
         let adata_gen = || AnnData::<H5>::new(&file).unwrap();
         test_iterator(|| adata_gen());
     })
+}
+
+#[test]
+fn test_io_h5_csc() {
+    let path = Path::new("/home/fangzq/github/anndata-rs/LLP.csc.h5ad");
+    let file = H5::open(path.to_path_buf()).unwrap();
+    let adata = AnnData::<H5>::open(file).unwrap();
+    // println!("adata csc_matrix: {:?}", &adata);
+    // // Automatical data type casting
+    // adata.x().get::<Array2<f64>>().unwrap().unwrap();
+    // adata.x().get::<ArrayData>().unwrap().unwrap()
+    let idxx = vec![0, 3, 5];
+    let gixx = s![.., idxx];
+    // let inner = adata.get_x().inner();
+    // let gex = inner.select::<CscMatrix<f32>, _>(gixx.as_ref()).unwrap();
+    let gex = adata.x().slice::<CscMatrix<f32>, _>(gixx).unwrap().unwrap();// CSCMatrix
+
+    // println!("{}", inner.dtype());
+    // println!("{}", inner.shape());
+    // let x = adata.x().get::<CscMatrix<f32>>().unwrap().unwrap();
+    println!("{:?}", gex.shape());
+    // let data = inner.data::<CsrMatrix<f32>>()?;
+    // println!("{:?}", data);
+
+
 }
